requirements.txt
#sudo apt-get install python3-tk
#sudo apt-get install libjpeg-dev zlib1g-dev
#sudo apt-get install python3-pip
#sudo pip3 install Pillow

pyzk
tinydb
requests
paho-mqtt
Pillow
rocketchat_API